[nitter] Add Media Gallery to Details View
Instead of rendering the images for a Nitter item within the description
in the details view, we are now rendering the images via the
"ItemMediaGallery" widget. To not render the images twice, we also added
the "disableImages" paramter to the "ItemDescription" widget, to not
render the images in the widget.