[core] Add Continuous Delivery for Desktop Apps (#6)
Add continuous delivery workflows for the desktop apps. While we want to
provide the desktop versions of FeedDeck though the different app stores
we still want to build them via GitHub Actions, so users can also test
the desktop apps before an official release.

This is also useful to build the binaries we can then download and
upload to the official stores, since we currently only have a macOS
system to build the apps.