[core] Add Continuous Delivery for Supabase (#7)
This commit adds the continuous delivery pipeline for Supabase, to
automatically push all database migrations and to deploy all Supabase
functions.

When a PR is merged to the "main" branch the pipeline is run for the
STAGE project. If a new tag is created the pipeline is run for the PROD
project.