[youtube] Add Missing URL in isYoutubeUrl Function
In the "isYoutubeUrl" function we are checking if the provided url is
related to YouTube. For that we checked if the url starts with
"https://www.youtube.com/" or "https://m.youtube.com/", but we also have
to check if the url starts with "https://youtube.com/" which is used
when a users shares a link to a channel via the YouTube app.

In this commit we add the missing "https://youtube.com/" option to the
"isYoutubeUrl" function.