[core] Fix Scrolling in List of Sources in Columns
Above the items of a column we display a list of sources, which can be
used by a user to only show items for a selected source. If a column
contains a lot of sources the list is vertical scrollable.

This vertical scrolling is not working on Linux, Windows and the Web and
only works on touch screens and on the magic trackpad for macOS, because
at the same time a user can also scroll horizontally.

To fix this we decided to move the list of sources out of the list of
items on macOS, Linux, Windows and on the web for large screens. This
way the vertical scroll can not conflict with the horizontal scroll.
This also means that the sources are always displayed on these
platforms.

For iOS, Android and on the web for small screens we do not changed the
behaviour. Since we assume that a user always has a touch screen for the
scrolling. We can also not change is because we do not have that much
screen space on these platforms and have to safe the space for the list
of items.