[core] Fix Vertical Scrolling
As reported in #25 the list of sources in a column can not be scrolled
vertically when there are a lot of sources which can not all be
displayed within the existing space.

This problem only occures when a user has no touch device or trackpad,
e.g. on Linux, Windows or the Web.

To fix this we changed the scroll behaviour in the app to allow
scrolling via drag gestures. For the reported case this means the list
of sources can be dragged to the right or left side to view all the
other sources which are hidden.