[podcast] Fix Audio Player for Windows and Linux
Instead of using the "just_audio_windows" and "just_audio_mpv" packages
for Windows and Linux, we are now using "just_audio_media_kit", so that
the audio player is working reliable on these platforms.

Note: We have to adjust the Flatpak, to include the dependencies which
are required on Linux.