[core] Fix Status Bar for iOS (#43)
The status bar wasn't visible on iOS when the application was launched.
This is now fixed so that a user can still see the current time, battery
life, etc. when the application is running.

This commit also updates the Podfile for iOS and macOS, which was not
done within #42, so that some dependencies were missing.