[core] Fix Border Radius for Drawer (#44)
This commit fixes the border radius for the drawer used in the large
deck layout. Until now the drawer used the border radius on all edges.
This is now fixed to only use the border radius on the right side of the
drawer. This is more aligned how the border radius is used in the side
sheet for the item details, where we also only use it on the left side.