[core] Fix Number of Loaded Items per Fetch
We always loaded 51 items per fetch instead of 50 items, because we used
a range of 0 to 50 which includes the item at position 0 and at position
50. This is now fixed by adjusting the range to 0 to 49, so that exactly
50 items are loaded per fetch request.

This commit also improves the number of loaded items shown in the column
header. If the status of the ItemsRepository is "loaded" we add a small
"+" symbol behind the number to indicat that there are more items, which
could be loaded. If the last page was already fetched (status ==
"loadedLast") then the "+" symbol isn't shown.