[core] Fix Blockquote Style (#148)
This commit fixes the style of blockquotes when we render the item
description as Markdown. Until now blockquotes had a weird blue
background, so that the text wasn't readable.

Now we are using the secondary color as background and we add a border on
the left site with the primary color.

We would also like to use italic as font family, but there is currently
a bug, so that the defined blockquote style is not applied. See
https://github.com/flutter/flutter/issues/81720