[core] Adjust Dependabot Configuration
Adjust the Dependabot configuration to not update the
`/supabase/email-templates` folder anymore. This is done, because
Dependabot couldn't handle the `package.json` and `package-lock.json`
file very well and always wants to delete some required stuff.

Since the folder only contains our email templates it is also not
critical if the dependencies are not up to date.